nvme-cli/spdk: handle IO commands with zero based blkcnt
The following IO related commands are zero based for the block count:
Read / Write / Compare / Write-Zeroes / Write-Uncor

Add the proper handling here for Read / Write / Compare commands
which is one based in SPDK for the buffer allocation and then
internally updated to zero based to follow the Spec.

The Write-Zeros / Write-Uncor commands are same in SPDK to follow
the Spec for zero based.

Change-Id: I5d3f0785d59614f5a8a1779b57a53ff230e906ac
Signed-off-by: CaoGang <gang.cao@intel.com>