Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Hi Seth,
> 
> Why in the patch, we add +1 for the strlen, is there any detailed reported error message

Yeah, GCC 8 (Which is the version fedora 28 ships with) complains about the source and destination buffers being the same size in calls to strncpy. The warning is that you could possibly have a source buffer without a null terminator in it which would cause the destination buffer to be out of compliance with c string standards.

Patch-set: 1
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>