nvme-print: fix gcc 8 error
After rebasing on release 1.6, the only remaining gcc 8 issue is this:
In the call to sprintf on line 2417,gcc complains that we could be
truncating c since an integer can be up to 10 characters long. In order
to fix this, I changed c to a uint8_t and increased the length of key to
24 (enough to accomodate three characters for 255 and a null
terminator).

Signed-off-by: Seth Howell <seth.howell@intel.com>